requirements.txt
pyrogram==2.0.106
httpx==0.23.3
# pymongo==4.9.2
python-dotenv==0.20.0
tgcrypto
# dnspython
# Flask==2.3.2
# gunicorn==20.1.0
motor==3.6.0
setuptools>=68.0.0
pytz==2023.3
